Refactor auth section components
